+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- The selected subtitle track label no longer includes the number of available tracks.
+
 ### Fixed
 
+- Long option names causing settings panels to overflow or setting labels to become unreadable.
 - Ad count in `AdCounterLabel` no longer increases incorrectly when used on mobile SDKs.
 
 ## [4.20.2] - 2026-09-03